WR343674 Block all access to jsmol.php
jsmol.php functions as a sort of cross-origin proxy with significantly more
access to Moodle than it should have: under the right conditions it can
leak the contents of PHP files including config.php, as well as attempt to
save files to the local file system for later exploitation. Logs do not
suggest it is currently being used, so simply break all the functionality
until we can sanitise it.

Change-Id: I155e384cae9d32a8b20c66031f286852538b1f14